Refactor reducers for screen sharing and session management
- Renamed reducer imports for consistency and clarity.
- Removed obsolete reducer files for screen sharing IDs and session management, streamlining the codebase.
- Ensured proper state management by consolidating reducer logic.
