+++ webapp/src/reducers.ts
@@ -6,7 +6,7 @@
 import {CallJobState, CallsConfig, CallsVersionInfo, LiveCaption, Reaction, UserSessionState} from '@mattermost/calls-common/lib/types';
 import {combineReducers} from 'redux';
 import {MAX_NUM_REACTIONS_IN_REACTION_STREAM} from 'src/constants';
-import {reducer as screenSharingIDs} from 'src/state/screen_sharing_ids/reducers';
+import {reducer as screenSharingIDs} from 'src/state/screen_sharing_ids/reducer';
 import {
     CALL_ENDED,
     UN_INITIALIZED,
@@ -15,7 +15,7 @@ import {
     USER_REACTED,
     USER_REACTED_TIMEOUT,
 } from 'src/state/session/action_types';
-import {reducer as sessions} from 'src/state/session/reducers';
+import {reducer as sessions} from 'src/state/session/reducer';
 import {
     CallsConfigDefault,
     CallsUserPreferences,
